% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPropsFunction.r
\name{calcProps}
\alias{calcProps}
\title{Calculates the proportion of probes (CpGs) with power > 0.8}
\usage{
calcProps(allCellRes)
}
\arguments{
\item{allCellRes}{- the list object returned from running either calcDiff() or calcSamples()}
}
\value{
- a dataframe where rows contain the proportion of probes with power > 0.8 for the range of mean differences or number of samples tested, and columns are each cell type
}
\description{
Calculates the proportion of probes (CpGs) with power > 0.8
}
\examples{
allProps <-calcProps(allSamples)
}
