% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPropsFunction.r
\name{calcProps}
\alias{calcProps}
\title{Calculates the proportion of probes (CpGs) with power > 0.8}
\usage{
calcProps(allCellRes)
}
\arguments{
\item{allCellRes}{- a list of matrices with one matrix per cell type returned from running either calcDiff() or calcSamples()}
}
\value{
- a dataframe containing the proportion of probes with power > 0.8 where rows are either mean differences or number of samples tested, and columns are each cell type
}
\description{
Calculates the proportion of probes (CpGs) with power > 0.8
}
