% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction.r
\name{plotPower}
\alias{plotPower}
\title{plot power curve}
\usage{
plotPower(resFile, calcType)
}
\arguments{
\item{resFile}{- the dataframe object output from running calcProps()}

\item{calcType}{- character string of the type of power calcualtion run. Either "samples" (if calcSamples() was used) or "difference" (if calcDiff() was used).}
}
\value{
- power curve plot
}
\description{
plot power curve
}
\examples{
myPlot <- plotPower(allProps, "samples")
}
